python_basics_course/04_llm_and_prompts/requirements.txt
# requirements.txt
openai>=0.27.0
python-dotenv>=0.19.0
tiktoken>=0.3.0
